camera/test_api_available.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Non-Permitted Application for Alarm API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="alarm_helper.js"/>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

"use strict";

steps = [setUp, api_available, tearDown];
SimpleTest.waitForExplicitFinish();
runAll(steps);

function api_available(next) {
  ok(navigator.mozCameras, "mozCamera api is present");
  next();
}

</script>
</pre>
</body>
</html>
